refactor(storybook) consolidate storybook and component types
