+++ packages/calcite-components/src/components/block-section/block-section.tsx
@@ -14,6 +14,7 @@ import {
 import { focusFirstTabbable, toAriaBoolean } from "../../utils/dom";
 import { isActivationKey } from "../../utils/key";
 import { connectLocalized, disconnectLocalized, LocalizedComponent } from "../../utils/locale";
+import { FlipContext } from "../interfaces";
 import {
   connectMessages,
   disconnectMessages,
@@ -21,7 +22,7 @@ import {
   T9nComponent,
   updateMessages,
 } from "../../utils/t9n";
-import { FlipContext, Status } from "../interfaces";
+import { Status } from "../interfaces";
 import {
   componentFocusable,
   LoadableComponent,
