+++ packages/calcite-components/src/components/carousel/interfaces.ts
@@ -1,2 +1,4 @@
+export type ArrowType = "inline" | "edge" | "none";
+
 /* When the 'autoplay' property of type 'boolean | string' is set to true, the value is "". */
 export type AutoplayType = boolean | "paused" | "";
