+++ packages/calcite-components/src/components/tile/tile.tsx
@@ -107,10 +107,7 @@ export class Tile implements InteractiveComponent, SelectableComponent {
    *
    * @internal
    */
-  @Prop({ reflect: true }) layout: Exclude<
-    Layout,
-    "grid" | "inline" | "center" | "auto" | "fixed" | "none" | "horizontal-single"
-  > = "horizontal";
+  @Prop({ reflect: true }) layout: Extract<Layout, "horizontal" | "vertical"> = "horizontal";
 
   /**
    * Specifies the size of the component.
